Change cr to nl.
When all this lived in the kernel, cr = '\r' was the line break.
I don't even remember why.
It was easier I guess.
Now in user space, newline = '\n' is king.
I switched over cr to nl in jencode.c and jxlate.c,
in a global substitute, and I hope this does more good than harm.