Read 345678 as digits, not millions thousands etc
This always worked in doubletalk, but apparently not espeak.
I'm sure this is just the beginning of things I will run into
that worked just fine on my hardware syngh, but not in espeak.
It will be quite a challenge to write a system that sounds sort of the same across all synths.