Add the -a option to the jupiter command line.
-a is for alphanumeric translation, like dates and times and fractions
and phone numbers and urls, and things you just can't do in a user dictionary.
This is good stuff,mostly, but only implemented in English,
so you should be able to turn it on and off.

Put more documentation into the README file,
including a list of all the atomic speech commands.